OUSD collateral swap fork tests Add Swapper approvals for DAI, USDT and USDC
