+++ contracts/deploy/070_flux_strategy.js
@@ -12,8 +12,6 @@ module.exports = deploymentWithGovernanceProposal(
     forceDeploy: false,
     // forceSkip: true,
     deployerIsProposer: true,
-    proposalId:
-      "91660566454734065741871109032962399777954404025908021614421255240824728026045",
   },
   async ({ ethers }) => {
     const { deployerAddr, timelockAddr } = await getNamedAccounts();
